SP - blacklisting some hosts at a early level
see https://github.com/georchestra/georchestra/pull/1704 ; the idea was
to avoid information disclosure in certain cases (e.g. in a AWS context,
do not allow to get URLs like http://ec2-metadata/ from the SP), but
actually these service are unlikely to be on a Class C address
(169.254.x.y are not Class C IPs), and the previous PR was also breaking
the feature allowing to fetch OGC servers in the same network but
outside of the geOrchestra solution.

This has been reported by Rennes-Métropole in the following (private)
issue:

https://github.com/camptocamp/georchestra-rennes-configuration/issues/619#issuecomment-328531487